pyproject.toml
[project]
name = "doc-obj-detect"
version = "0.1.0"
description = "Document object detection with vision transformers and knowledge distillation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "transformers>=4.46.0",
    "datasets>=3.1.0",
    "timm>=1.0.0",
    "tensorboard>=2.18.0",
    "pyyaml>=6.0",
    "pillow>=11.0.0",
    "pycocotools>=2.0.8",
    "opencv-python>=4.10.0",
    "numpy>=2.0.0",
    "tqdm>=4.66.0",
    "pydantic>=2.0.0",
    "accelerate>=0.26.0",
    "matplotlib>=3.9.0",
    "albumentationsx>=2.0.13",
    "torchmetrics>=1.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
]
cu128 = [
    "torch>=2.5.0",
    "torchvision>=0.20.0",
]

[project.scripts]
doc-obj-detect = "doc_obj_detect.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/doc_obj_detect"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src/doc_obj_detect"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "extra == 'cu128'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "extra == 'cu128'" },
]
